% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_ursus}
\alias{barplot_ursus}
\title{Get barplot of the number of ecoregions per species}
\usage{
barplot_ursus(data_ursus_eco_reg)
}
\arguments{
\item{data_ursus_eco_reg}{dataset with both bears and eco_region data}
}
\value{

}
\description{
Get barplot of the number of ecoregions per species
}
