% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{clean_pantheria}
\alias{clean_pantheria}
\title{Get cleaned data pantheria}
\usage{
clean_pantheria(datapantheria)
}
\arguments{
\item{datapantheria}{data from pantheria database obtain with the function datapantheria.R}
}
\value{

}
\description{
Get cleaned data pantheria
}
