% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{combine_all_ursus}
\alias{combine_all_ursus}
\title{Combination of all info}
\usage{
combine_all_ursus(ursus_eco_realm, ursus_eco_biomes, eco_ursus, ursus_list)
}
\arguments{
\item{ursus_list}{}
}
\value{
ursus_all_infos a tibble containing all info on Ursidae species
}
\description{
Combination of all info
}
