% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_ecoregion}
\alias{data_ecoregion}
\title{Get data ecoregion}
\usage{
data_ecoregion()

data_ecoregion()

data_ecoregion()
}
\value{
data.frame with info on ecoregion

data.frame with info on mammals

data.frame with info on species
}
\description{
Get data ecoregion

Get data mammals

Get data species
}
