% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_ecoregions_mammals}
\alias{data_ecoregions_mammals}
\title{Get mammal species <-> ecoregions data}
\usage{
data_ecoregions_mammals()
}
\value{
data.frame with infos on which mammals in ecoregions
}
\description{
Get mammal species <-> ecoregions data
}
