% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_ecoregions_species}
\alias{data_ecoregions_species}
\title{Get ecoregions and species data}
\usage{
data_ecoregions_species()
}
\value{
tibble with info on which mammal species is found in which ecoregions
}
\description{
Get ecoregions and species data
}
