% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_mammals_ecoregion}
\alias{data_mammals_ecoregion}
\title{Get mammal <-> ecoregion data}
\usage{
data_mammals_ecoregion()
}
\value{
data.frame on which mammals in ecoregions
}
\description{
Get mammal <-> ecoregion data
}
