% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_pantheria}
\alias{data_pantheria}
\title{Import pantheria data}
\usage{
data_pantheria()
}
\value{
A 2 columns tibble linking species IDs to ecoregions IDs
}
\description{
Import pantheria data
}
