% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_panteria.R
\name{data_plot_families}
\alias{data_plot_families}
\title{Compute the number of individuals per family}
\usage{
data_plot_families(pantheria)
}
\arguments{
\item{pantheria}{}
}
\value{
a tibble with the number of individuals per family
}
\description{
Compute the number of individuals per family
}
