% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_panteria.R
\name{data_plot_littersize_longevity}
\alias{data_plot_littersize_longevity}
\title{Prepare data to plot litter size versus longevity}
\usage{
data_plot_littersize_longevity(pantheria)
}
\arguments{
\item{pantheria}{}
}
\value{
a tibble
}
\description{
Prepare data to plot litter size versus longevity
}
