% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pantheria.R
\name{explo_data}
\alias{explo_data}
\title{Explore the data}
\usage{
explo_data(tidy_pantheria)
}
\value{
a tibble with mean home range by family? Standard deviation? The sample size?
}
\description{
Explore the data
}
