% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{extract_ursidae}
\alias{extract_ursidae}
\title{To extract usrus data}
\usage{
extract_ursidae(sp)
}
\arguments{
\item{sp}{The data containing all species i.e. sp_list obtain with data_mammals() function}
}
\value{

}
\description{
To extract usrus data
}
