% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gestation_longevity}
\alias{gestation_longevity}
\title{Get Gestation~longevity figure}
\usage{
gestation_longevity(data)
}
\arguments{
\item{data}{cleaned pantheria data set used to draw the figure}
}
\value{

}
\description{
Get Gestation~longevity figure
}
