% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{get_biome_nb}
\alias{get_biome_nb}
\title{Number of biomes where each Ursidae species is found}
\usage{
get_biome_nb(ursus_eco)
}
\arguments{
\item{ursus_eco}{}
}
\value{
ursus_eco_biomes a tibble counting the number of biomes containing Ursidae species
}
\description{
Number of biomes where each Ursidae species is found
}
