% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{get_ecoregions_nb}
\alias{get_ecoregions_nb}
\title{Number of ecoregions where each Ursidae species is found}
\usage{
get_ecoregions_nb(ursus_eco)
}
\arguments{
\item{ursus_eco}{}
}
\value{
ursus_ecoregions a tibble counting the number of ecoregions containing Ursidae species
}
\description{
Number of ecoregions where each Ursidae species is found
}
