% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_nb_ecoreg}
\alias{get_nb_ecoreg}
\title{Get the number of ecoregion per ursidae species}
\usage{
get_nb_ecoreg(data)
}
\arguments{
\item{data}{ursus_eco obained with join_dataset() function}
}
\value{

}
\description{
Get the number of ecoregion per ursidae species
}
