% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{get_necoregions}
\alias{get_necoregions}
\title{Ecoregion ursidae}
\usage{
get_necoregions()
}
\value{
number of different ecoregions in which we found each of the 7 species
}
\description{
Ecoregion ursidae
}
