% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{get_realm_nb}
\alias{get_realm_nb}
\title{Number of realms where each Ursidae species is found}
\usage{
get_realm_nb(ursus_eco)
}
\arguments{
\item{ursus_eco}{}
}
\value{
ursus_realm a tibble counting the number of realms containing Ursidae species
}
\description{
Number of realms where each Ursidae species is found
}
