% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pantheria.R
\name{Graph1}
\alias{Graph1}
\title{Bar chart of the number of entries per family}
\usage{
Graph1(tidy_pantheria)
}
\value{
A ggplot bar chart of the number of entries per family
}
\description{
Bar chart of the number of entries per family
}
