% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pantheria.R
\name{Graph2}
\alias{Graph2}
\title{Scatter plot of litter size as a function of longevity}
\usage{
Graph2(tidy_pantheria)
}
\value{
A ggplot scatter plot of litter size as a function of longevity
}
\description{
Scatter plot of litter size as a function of longevity
}
