% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{join_data}
\alias{join_data}
\title{Join dataframes}
\usage{
join_data(data_ursus, data_speco, data_eco)
}
\arguments{
\item{data_ursus}{The data containing ursidae species}

\item{data_speco}{The data containing species and ecoregions}

\item{data_eco}{The data containing ecoregion info}
}
\value{
A tibble containing ursidae with ecoregion info
}
\description{
Join dataframes
}
