% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{join_dataset}
\alias{join_dataset}
\title{Join the three data set ursus species, mammals and ecoregion and ecoregion}
\usage{
join_dataset(data1, data2, data3)
}
\arguments{
\item{data1}{ursus data obtained with extract_ursidae() function}

\item{data2}{sp_eco data set linking species and eco region , obtained with data_mammals_ecoregions()}

\item{data3}{eco_list list of eco_regions obtained with data_ecoregion() function}
}
\value{

}
\description{
Join the three data set ursus species, mammals and ecoregion and ecoregion
}
