% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{joints_ursus_eco}
\alias{joints_ursus_eco}
\title{Joint in order to have ecoregions info}
\usage{
joints_ursus_eco(ursus_list, sp_eco, eco_list)
}
\arguments{
\item{eco_list}{}
}
\value{
joints_ursus, a tibble containing infos on Ursus and ecoregions
}
\description{
Joint in order to have ecoregions info
}
