% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_panteria.R
\name{plot_families}
\alias{plot_families}
\title{Barplot of the number of individuals per family}
\usage{
plot_families(pantheria_plot_families)
}
\arguments{
\item{pantheria_plot_families}{}
}
\value{
a barplot of the number of individuals per family
}
\description{
Barplot of the number of individuals per family
}
