% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_panteria.R
\name{plot_littersize_longevity}
\alias{plot_littersize_longevity}
\title{Scatter plot of litter size versus longevity}
\usage{
plot_littersize_longevity(data_plot_littersize_longevity)
}
\arguments{
\item{pantheria_plot_littersize_longevity}{}
}
\value{
a scatter plot of litter size versus longevity
}
\description{
Scatter plot of litter size versus longevity
}
