% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plot}
\alias{save_plot}
\title{Plot gestation}
\usage{
save_plot(plot, name, ext = "png", width = 8, height = 6, ...)
}
\arguments{
\item{plot}{ggplot2 object}

\item{name}{character}

\item{ext}{png, pdf, tiff,... (default is png)}

\item{...}{}
}
\value{

}
\description{
Plot gestation
}
\examples{

}
