% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ursidae.R
\name{select_ursus}
\alias{select_ursus}
\title{Selection of Ursidae only}
\usage{
select_ursus(sp_list)
}
\arguments{
\item{sp_list}{}
}
\value{
ursus_list the list of all Ursidae species
}
\description{
Selection of Ursidae only
}
