% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pantheria.R
\name{tidy_pantheria}
\alias{tidy_pantheria}
\title{Tidy the pantheria dataset}
\usage{
tidy_pantheria(pantheria)
}
\arguments{
\item{pantheria}{}
}
\value{
tidy pantheria tibble
}
\description{
Tidy the pantheria dataset
}
