% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_mammals_ecoregion}
\alias{data_mammals_ecoregion}
\title{Get mammals ecoregions data}
\usage{
data_mammals_ecoregion()
}
\value{
data.frame with info on which mammals in ecoregions
}
\description{
Get mammals ecoregions data
}
