% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_mammals_ecoregions}
\alias{data_mammals_ecoregions}
\title{get mammal <-> ecoregions data  data}
\usage{
data_mammals_ecoregions()
}
\value{

}
\description{
get mammal <-> ecoregions data  data
}
\examples{

}
