% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_mammals_ecoregions}
\alias{data_mammals_ecoregions}
\title{Get mammals <-> ecoregions data}
\usage{
data_mammals_ecoregions()
}
\value{
data.frame with info on which mammals species in ecoregion
}
\description{
Get mammals <-> ecoregions data
}
