% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pantheria.R, R/load.R
\name{data_pantheria}
\alias{data_pantheria}
\title{Get pantheria data}
\usage{
data_pantheria()

data_pantheria()
}
\value{
data.frame with info on ecoregions

A 2 columns tibble linking species IDs to ecoregions IDs
}
\description{
Get pantheria data

Import pantheria data
}
