% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{extract_ursidae}
\alias{extract_ursidae}
\title{Exctract Ursidae species}
\usage{
extract_ursidae(sp)
}
\value{
a tibble with only info on Ursidae species
}
\description{
Exctract Ursidae species
}
