% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{extract_ursidae}
\alias{extract_ursidae}
\title{Extract ursidae}
\usage{
extract_ursidae(spdata)
}
\arguments{
\item{spdata}{The data containing all species}
}
\value{
A tibble only containing ursidae
}
\description{
Extract ursidae
}
