% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_necoregions}
\alias{get_necoregions}
\title{Get number of ecoregions}
\usage{
get_necoregions(data)
}
\arguments{
\item{data}{The joined tibble containing Ursidae species and ecoregion info}
}
\value{
A tibble with number of ecoregions per species of family Ursidae
}
\description{
Get number of ecoregions
}
