% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{tidy_pantheria}
\alias{tidy_pantheria}
\title{Tidy data pantheria}
\usage{
tidy_pantheria(data)
}
\arguments{
\item{data}{The raw pantheria dataset}
}
\value{
A cleaned tibble
}
\description{
Tidy data pantheria
}
